XGB Neural Nets Draft
Adding integration of neural networks with XGBoost models (presuming
GPU). A first pass is made through neural networks, which also export
the last dataframe they have before the final layer. This is then
imported into XGBoost functions and treated as the data input. This is
not functional yet, errors expected.